Remove TraverseSafariWebAreaElementService in favor of an if statement
